#15 Refactor a bit
