+++ platonic/__init__.py
@@ -1,3 +1,4 @@
 from platonic.type_args import generic_type_args
 from platonic.const import const
 from platonic.mapping import PlatonicMapping
+from platonic.memoize import memoize
